Changed physics engine so gravity is less harsh. vy is now a float and it is no longer required to add vy to y after setting vy. There is a new define for JUMP_VELOCITY. Velocity is no longer upside-down (Positive velocity = Up).

+++ MineDS/source/mobs/baseMob.h
@@ -9,7 +9,7 @@ class baseMob
 public:
 	int x;
 	int y;
-	int vy;
+	float vy;
 	int vx;
 	int sx;
 	int sy;
